+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [4.0.0] - Unreleased
+## [4.0.0] - 2026-08-19
 
 ### Changed